Extract createDialogHeading to common mixin
