+++ client/components/repository/common/ActivityOptions/common.js
@@ -1,9 +1,12 @@
 import { mapGetters, mapMutations } from 'vuex';
 import find from 'lodash/find';
 import get from 'lodash/get';
+import insertActions from '@/utils/insertActions';
 import { isEditable } from 'shared/activities';
 import selectActivity from '@/components/repository/common/selectActivity';
 import uniqBy from 'lodash/uniqBy';
+
+const { ADD_AFTER, ADD_BEFORE, ADD_INTO } = insertActions;
 
 export default {
   mixins: [selectActivity],
@@ -23,6 +26,14 @@ export default {
       const config = find(this.structure, { type: this.activity.type });
       const subLevels = get(config, 'subLevels', []);
       return this.structure.filter(it => subLevels.includes(it.type));
+    },
+    createDialogHeading: ({ action }) => {
+      const heading = {
+        [ADD_BEFORE]: 'Add above',
+        [ADD_AFTER]: 'Add below',
+        [ADD_INTO]: 'Add into'
+      };
+      return heading[action];
     }
   },
   methods: {
